+++ common/src/main/java/com/sunekaer/sdrp/SDRP.java
@@ -6,6 +6,7 @@ import com.sunekaer.sdrp.config.SDRPConfig;
 import com.sunekaer.sdrp.discord.RPClient;
 import com.sunekaer.sdrp.discord.State;
 //import com.sunekaer.sdrp.integration.kubejs.SDRPKubeJSIntegration;
+import com.sunekaer.sdrp.integration.kubejs.SDRPKubeJSIntegration;
 import dev.architectury.event.EventResult;
 import dev.architectury.event.events.client.ClientGuiEvent;
 import dev.architectury.event.events.client.ClientLifecycleEvent;
@@ -44,6 +45,10 @@ public class SDRP {
 
         EntityEvent.ADD.register(SDRP::clientJoinEvent);
         ClientGuiEvent.INIT_POST.register(SDRP::screenEvent);
+
+        if (Platform.isModLoaded("kubejs")) {
+            SDRPKubeJSIntegration.setup();
+        }
     }
 
     /**