+++ common/src/main/java/com/sunekaer/sdrp/discord/RPClient.java
@@ -89,6 +89,9 @@ public class RPClient {
         }
 
         this.client.sendRichPresence(state);
+        if (SDRP.config.logState) {
+            LOGGER.info("Sent state to discord: {}", state.toJson().toString());
+        }
     }
 
     /**