+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# Changelog
-Merge pull request #34  
-Updated some lang keys.
+## 4.0.3
+
+### Added
+
+- KubeJS integration back in with custom events
+
+### Changed
+
+- Improved the readme a bunch!